+++ ReClass/CNodeHex8.h
@@ -9,7 +9,7 @@ public:
 
 	virtual void Update( HotSpot& Spot );
 
-	virtual int GetMemorySize( ) { return 1; }
+	virtual ULONG GetMemorySize( ) { return 1; }
 
 	virtual int Draw( ViewInfo& View, int x, int y );
 };
