+++ ReClass/CNodeInt64.h
@@ -9,7 +9,7 @@ public:
 
 	virtual void Update( HotSpot& Spot );
 
-	virtual int GetMemorySize( ) { return sizeof( __int64 ); }
+	virtual ULONG GetMemorySize( ) { return sizeof( __int64 ); }
 
 	virtual int Draw( ViewInfo& View, int x, int y );
 };
