+++ ReClass/CNodeMatrix.h
@@ -7,7 +7,7 @@ public:
 
 	virtual void Update( HotSpot& Spot );
 
-	virtual int GetMemorySize( ) { return 4 * 4 * sizeof( float ); }
+	virtual ULONG GetMemorySize( ) { return 4 * 4 * sizeof( float ); }
 
 	virtual int Draw( ViewInfo& View, int x, int y );
 };
